pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitventory"
version = "0.1.0"
description = "Modular inventory service linking code repositories to cloud deployments"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "sqlalchemy>=2.0",
    "PyGithub>=2.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
gitventory = "gitventory.cli:main"

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
dev = [
    "pytest>=7",
    "pytest-mock",
    "responses",
    "black",
    "mypy",
    "types-PyYAML",
]

[tool.hatch.build.targets.wheel]
packages = ["gitventory"]
